# Simplify/Optimize HandlerTree for Constant Routes