@kitlangton 2022-09-23
```scala
// Original
//
// [info] ApiBenchmark.benchmarkBroadTapirHttp4s      thrpt    2  115.028          ops/s
// [info] ApiBenchmark.benchmarkBroadZioApi           thrpt    2  751.619          ops/s
// [info] ApiBenchmark.benchmarkBroadZioCollect       thrpt    2  126.598          ops/s
//
// [info] ApiBenchmark.benchmarkDeepPathTapirHttp4s   thrpt    2   65.697          ops/s
// [info] ApiBenchmark.benchmarkDeepPathZioApi        thrpt    2  565.441          ops/s
// [info] ApiBenchmark.benchmarkDeepPathZioCollect    thrpt    2  956.824          ops/s
//
// [info] ApiBenchmark.benchmarkSmallDataTapirHttp4s  thrpt    2   92.525          ops/s
// [info] ApiBenchmark.benchmarkSmallDataZioApi       thrpt    2  277.247          ops/s
// [info] ApiBenchmark.benchmarkSmallDataZioCollect   thrpt    2  703.899          ops/s
//
// After Branch optimization
//
// [info] ApiBenchmark.benchmarkBroadTapirHttp4s      thrpt    2  118.398          ops/s
// [info] ApiBenchmark.benchmarkBroadZioApi           thrpt    2  979.155          ops/s (up from 751)
// [info] ApiBenchmark.benchmarkBroadZioCollect       thrpt    2  122.202          ops/s
//
// [info] ApiBenchmark.benchmarkDeepPathTapirHttp4s   thrpt    2   65.647          ops/s
// [info] ApiBenchmark.benchmarkDeepPathZioApi        thrpt    2  563.393          ops/s
// [info] ApiBenchmark.benchmarkDeepPathZioCollect    thrpt    2  977.179          ops/s
//
// [info] ApiBenchmark.benchmarkSmallDataTapirHttp4s  thrpt    2   91.951          ops/s
// [info] ApiBenchmark.benchmarkSmallDataZioApi       thrpt    2  264.240          ops/s
// [info] ApiBenchmark.benchmarkSmallDataZioCollect   thrpt    2  698.570          ops/s
```

Here are the results for the difference between looking up constant paths in a map vs. hunting through a list of them. It looks like it earned us around +200 ops.

I tried out some other micro-optimizations for fun and also to help build my intuition for what makes a dent. Some of those included:

- Making the Chunk of results that we're building into an Array
- Pre-optimizing all Handlers into APIServers, removing the `Memoization`, and moving the logic that happens after `lookup` into its body.

But lo and behold, the code got uglier and there wasn't much of a bump. Maybe ~10/20 ops? But it didn't feel worth sacrificing the readability for such a small (potentially random) gain.