@kitlangton 2022-09-23
Only encoding the `Output` into a `Body` if it's not `Unit` has sped up a couple of cases quite significantly. 

```scala
// After `hasBody` check
//
// [info] ApiBenchmark.benchmarkBroadTapirHttp4s      thrpt    2   113.092          ops/s
// [info] ApiBenchmark.benchmarkBroadZioApi           thrpt    2  1745.498          ops/s (from 751)
// [info] ApiBenchmark.benchmarkBroadZioCollect       thrpt    2   125.484          ops/s
//
// [info] ApiBenchmark.benchmarkDeepPathTapirHttp4s   thrpt    2    66.033          ops/s
// [info] ApiBenchmark.benchmarkDeepPathZioApi        thrpt    2   781.366          ops/s (from 563)
// [info] ApiBenchmark.benchmarkDeepPathZioCollect    thrpt    2   944.260          ops/s
//
// [info] ApiBenchmark.benchmarkSmallDataTapirHttp4s  thrpt    2    87.859          ops/s
// [info] ApiBenchmark.benchmarkSmallDataZioApi       thrpt    2   267.333          ops/s (the same because it encodes)
// [info] ApiBenchmark.benchmarkSmallDataZioCollect   thrpt    2   701.566          ops/s
```

Obviously, many APIs _will_ have some kind of body. But it's good to know that's the cause of some delay. It would appear that using `Schema` to decode into `Json` might be a good target for optimization, as the `smallDataCollect` benchmark uses `zip-json` directly. I'll switch that over to doing the same and see if they even out.